requirements.txt
flask
facenet-pytorch
torch
numpy
pillow
